+++ src/global.cpp
@@ -7,6 +7,4 @@ int pass = 0, IsLabelNotFound = 0;
 
 aint CurrentGlobalLine = 0, CurrentLocalLine = 0, CompiledCurrentLine = 0;
 
-int ErrorCount = 0;
-
 stack<RepeatInfo> RepeatStack; /* added */