+++ src/message_if.cpp
@@ -1,0 +1,4 @@
+namespace msg {
+    int WarningCount = 0;
+    int ErrorCount = 0;
+}