+++ src/parser/message.h
@@ -4,27 +4,48 @@
 #include <string>
 #include <vector>
 #include <tao/pegtl/parse_error.hpp>
+#include <tao/pegtl/position.hpp>
 
-#include <message_if.h>
+#include "../message_if.h"
+#include "../message.h"
 
 namespace parser {
 
     template <typename MP>
-    class MsgPrinter {
+    class MessagePrinter {
     public:
+
         static void note(const tao::pegtl::parse_error &E) {
-            MP::note(getPos(E), E.what());
+            P::note(getPos(E), E.what());
         }
+        static void note(tao::TAO_PEGTL_NAMESPACE::position P, const std::string &Msg) {
+            P::note(getPos(P), Msg);
+        }
+
         static void warn(const tao::pegtl::parse_error &E) {
-            MP::warn(getPos(E), E.what());
+            P::warn(getPos(E), E.what());
         }
+        static void warn(tao::TAO_PEGTL_NAMESPACE::position P, const std::string &Msg) {
+            P::warn(getPos(P), Msg);
+        }
+
         static void err(const tao::pegtl::parse_error &E) {
-            MP::err(getPos(E), E.what());
+            P::err(getPos(E), E.what());
         }
+        static void err(tao::TAO_PEGTL_NAMESPACE::position P, const std::string &Msg) {
+            P::err(getPos(P), Msg);
+        }
+
         [[noreturn]] static void fatal(const tao::pegtl::parse_error &E) {
-            MP::fatal(getPos(E), E.what());
+            P::fatal(getPos(E), E.what());
+        }
+        [[noreturn]] static void fatal(tao::TAO_PEGTL_NAMESPACE::position P, const std::string &Msg) {
+            P::fatal(getPos(P), Msg);
         }
     private:
+
+        using P = msg::IMessagePrinter<MP>;
+
         static msg::Position getPos(const tao::pegtl::parse_error &E) {
             const auto &Pos = E.positions.back();
             return {
@@ -34,7 +55,17 @@ namespace parser {
             };
         }
 
+        static msg::Position getPos(const tao::TAO_PEGTL_NAMESPACE::position &P) {
+            return {
+                P.source,
+                P.line,
+                P.byte_in_line + 1
+            };
+        }
+
     };
+
+    using M = MessagePrinter<msg::MessagePrinter>;
 
 } // namespace parser