+++ ChangeLog.md
@@ -1,8 +1,11 @@
 # Change Log
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
-## 2019-XX-YY
+## 2021-XX-YY
 - Version 20190306.1++WiP
+
+### Changed
+- Redefinition of `DEFINE`s is now a warning, not an error.
 
 ### Fixed
 - `END` was not terminating parsing if there were more lines in the buffer